refactored & modularized device details to ease composability
Changelog: None
